+++ src/Module.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace mix8872\filesAttacher;
 
 use yii\filters\AccessControl;
@@ -9,7 +10,13 @@ use yii\base\InvalidConfigException;
  */
 class Module extends \yii\base\Module
 {
-	public $parameters;
+    public $parameters;
+
+    // defaults
+    const FILES_NAME_BY = 'random';
+    const IMG_PROCESS_DRIVER = 'gd';
+    const SIZES_NAME_BY = 'size';
+    const SIZES_NAME_TEMPLATE = '%s';
 
     /**
      * @inheritdoc
@@ -20,19 +27,24 @@ class Module extends \yii\base\Module
         $this->controllerNamespace = 'mix8872\filesAttacher\controllers';
         $this->setViewPath('@vendor/mix8872/files-attacher/src/views');
         $this->registerTranslations();
-        
+
+        if (!isset($this->parameters['filesNameBy'])) {
+            $this->parameters['filesNameBy'] = self::FILES_NAME_BY;
+        }
+
+        if (!isset($this->parameters['imgProcessDriver'])) {
+            $this->parameters['imgProcessDriver'] = self::IMG_PROCESS_DRIVER;
+        }
+
         if (!isset($this->parameters['sizesNameBy'])) {
-            $this->parameters['sizesNameBy'] = 'size';
+            $this->parameters['sizesNameBy'] = self::SIZES_NAME_BY;
+        }
+
+        if (!isset($this->parameters['sizesNameTemplate'])) {
+            $this->parameters['sizesNameTemplate'] = self::SIZES_NAME_TEMPLATE;
         }
 
         if ($this->parameters['sizesNameBy'] === 'template'
-            && (
-                !isset($this->parameters['sizesNameTemplate'])
-                || empty($this->parameters['sizesNameTemplate'])
-            )
-        ) {
-            return new InvalidConfigException('Parameter "sizesNameBy" set to "template", but template is not defined or empty! Please define "sizesNameTemplate" parameter.');
-        } elseif ($this->parameters['sizesNameBy'] === 'template'
             && preg_match('/(%k|%s)/u', $this->parameters['sizesNameTemplate']) === false
         ) {
             return new InvalidConfigException('Parameter "sizesNameBy" set to "template", but template does not meet the requirements! Template must contain at least one of the characters "%k" or/and "%s".');
@@ -44,11 +56,11 @@ class Module extends \yii\base\Module
      */
     public function registerTranslations()
     {
-         \Yii::$app->i18n->translations['files'] = [
+        \Yii::$app->i18n->translations['files'] = [
             'class' => 'yii\i18n\PhpMessageSource',
             'sourceLanguage' => 'ru-RU',
             'basePath' => '@vendor/mix8872/files-attacher/src/messages',
-            ];
- 
+        ];
+
     }
 }